+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added hyprsplit compatibility for hy3:movetoworkspace
 - Added support for special workspaces
 - Added monitor switching support to hy3:movefocus
+- Added support for gapsin and gapsout workspace rules
 
 # hl0.44.0 and before